feat: 新增 deleteBrokerStockData 方法、調整 getBrokers 回應欄位

+++ backend2/src/main/java/com/cmoney/backend2/brokerdatatransmission/service/api/brokerstockdata/get/StockInfo.kt
@@ -7,7 +7,7 @@ class StockInfo(
     @SerializedName("tradeType")
     val tradeType: TradeType?,
     @SerializedName("amount")
-    val amount: Int?,
+    val amount: Long?,
     @SerializedName("tradeTotalCost")
     val tradeTotalCost: Double?
 )
